feat: reorganize hooks structure by moving useComparator, useDiffLineProps, and useDownloadJSON to hooks directory
