+++ src/webapp/components/comparator/hooks/useDownloadJSON.tsx
